+++ jodd-decora/src/test/java/jodd/decora/DecoraManagerTest.java
@@ -25,24 +25,22 @@
 
 package jodd.decora;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.powermock.api.mockito.PowerMockito.mock;
-import static org.powermock.reflect.Whitebox.getInternalState;
-import static org.powermock.reflect.Whitebox.setInternalState;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.junit.Before;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.mock;
 
 public class DecoraManagerTest {
 
 	private DecoraManager decoraManager;
 
-	@Before
+	@BeforeEach
 	public void setUp() {
 		decoraManager = new DecoraManager();
 	}
@@ -50,19 +48,10 @@ public class DecoraManagerTest {
 	@Test
 	public final void testIsDecorateErrorPages() {
 		// when
-		setInternalState(decoraManager, "decorateErrorPages", true);
-
-		// then
-		assertTrue("DecorateErrorPages should be true.", decoraManager.isDecorateErrorPages());
-	}
-
-	@Test
-	public final void testSetDecorateErrorPages() {
-		// when
 		decoraManager.setDecorateErrorPages(true);
 
 		// then
-		assertTrue("DecorateErrorPages should be true.", (boolean) getInternalState(decoraManager, "decorateErrorPages"));
+		assertTrue(decoraManager.isDecorateErrorPages());
 	}
 
 	@Test
@@ -71,7 +60,7 @@ public class DecoraManagerTest {
 		HttpServletRequest httpServletRequestMock = mock(HttpServletRequest.class);
 
 		// then
-		assertTrue("DecorateRequest function always returns true.", decoraManager.decorateRequest(httpServletRequestMock));
+		assertTrue(decoraManager.decorateRequest(httpServletRequestMock));
 	}
 
 	@Test
@@ -80,7 +69,7 @@ public class DecoraManagerTest {
 		String testString = "TEST";
 
 		// then
-		assertTrue("DecorateContentType function always returns true.", decoraManager.decorateContentType(testString, testString, testString));
+		assertTrue(decoraManager.decorateContentType(testString, testString, testString));
 	}
 
 	@Test
@@ -89,37 +78,37 @@ public class DecoraManagerTest {
 		int statusCode = 200;
 
 		// then
-		assertTrue("Returns <code>true</code> for status code 200", decoraManager.decorateStatusCode(statusCode));
+		assertTrue(decoraManager.decorateStatusCode(statusCode));
 	}
 
 	@Test
 	public final void testDecorateStatusCode2() {
 		// when
 		int statusCode = 300;
-		setInternalState(decoraManager, "decorateErrorPages", true);
+		decoraManager.setDecorateErrorPages(true);
 
 		// then
-		assertFalse("Returns <code>false</code> for status code 300", decoraManager.decorateStatusCode(statusCode));
+		assertFalse(decoraManager.decorateStatusCode(statusCode));
 	}
 
 	@Test
 	public final void testDecorateStatusCode3() {
 		// when
 		int statusCode = 404;
-		setInternalState(decoraManager, "decorateErrorPages", false);
+		decoraManager.setDecorateErrorPages(false);
 
 		// then
-		assertFalse("Returns <code>false</code> for status code 404", decoraManager.decorateStatusCode(statusCode));
+		assertFalse(decoraManager.decorateStatusCode(statusCode));
 	}
 
 	@Test
 	public final void testDecorateStatusCode4() {
 		// when
 		int statusCode = 404;
-		setInternalState(decoraManager, "decorateErrorPages", true);
+		decoraManager.setDecorateErrorPages(true);
 
 		// then
-		assertTrue("For error pages (status code {@literal >=} 400) should return true", decoraManager.decorateStatusCode(statusCode));
+		assertTrue(decoraManager.decorateStatusCode(statusCode));
 	}
 
 	@Test
@@ -132,7 +121,7 @@ public class DecoraManagerTest {
 		String result = decoraManager.resolveDecorator(httpServletRequestMock, actionPath);
 
 		// then
-		assertNull("If decorator is not found, returns <code>null</code>.", result);
+		assertNull(result);
 	}
 
 	@Test
@@ -145,7 +134,7 @@ public class DecoraManagerTest {
 		String result = decoraManager.resolveDecorator(httpServletRequestMock, actionPath);
 
 		// then
-		assertEquals("Result value must be equal to default decorator path.", DecoraManager.DEFAULT_DECORATOR, result);
+		assertEquals(DecoraManager.DEFAULT_DECORATOR, result);
 	}
 
 }