+++ jodd-decora/src/test/java/jodd/decora/DecoraResponseWrapperTest.java
@@ -25,20 +25,18 @@
 
 package jodd.decora;
 
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.powermock.api.mockito.PowerMockito.mock;
-import static org.powermock.api.mockito.PowerMockito.when;
-import static org.powermock.reflect.Whitebox.getInternalState;
+import jodd.servlet.wrapper.LastModifiedData;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.junit.Before;
-import org.junit.Test;
-
-import jodd.servlet.wrapper.LastModifiedData;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 public class DecoraResponseWrapperTest {
 
@@ -48,7 +46,7 @@ public class DecoraResponseWrapperTest {
 	private LastModifiedData lastModifiedData;
 	private DecoraManager decoraManager;
 
-	@Before
+	@BeforeEach
 	public void setUp() {
 		originalRequest = mock(HttpServletRequest.class);
 		originalResponse = mock(HttpServletResponse.class);
@@ -62,10 +60,10 @@ public class DecoraResponseWrapperTest {
 		decoraResponseWrapper = new DecoraResponseWrapper(originalRequest, originalResponse, lastModifiedData, decoraManager);
 
 		// then
-		assertEquals("Parameter should be set.", originalRequest, getInternalState(decoraResponseWrapper, "request"));
-		assertEquals("Parameter should be set.", originalResponse, getInternalState(decoraResponseWrapper, "response"));
-		assertEquals("Parameter should be set.", lastModifiedData, getInternalState(decoraResponseWrapper, "lastModifiedData"));
-		assertEquals("Parameter should be set.", decoraManager, getInternalState(decoraResponseWrapper, "decoraManager"));
+		assertEquals(originalRequest, decoraResponseWrapper.request);
+		assertEquals(originalResponse, decoraResponseWrapper.response);
+		assertEquals(lastModifiedData, decoraResponseWrapper.getLastModifiedData());
+		assertEquals(decoraManager, decoraResponseWrapper.decoraManager);
 	}
 
 	@Test