+++ jodd-decora/src/test/java/jodd/decora/parser/DecoratorTagVisitorTest.java
@@ -25,44 +25,40 @@
 
 package jodd.decora.parser;
 
-import static org.junit.Assert.fail;
-import static org.powermock.reflect.Whitebox.invokeMethod;
-import static org.powermock.reflect.Whitebox.setInternalState;
-
-import org.junit.Before;
-import org.junit.Test;
-
 import jodd.decora.DecoraException;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class DecoratorTagVisitorTest {
 
 	private DecoratorTagVisitor decoraTagVisitor;
 
-	@Before
+	@BeforeEach
 	public void setUp() {
 		decoraTagVisitor = new DecoratorTagVisitor();
 	}
 
-	@Test(expected = DecoraException.class)
+	@Test
 	public final void testCheckNestedDecoraTagsDecoraTagNameNotNull() throws Exception {
 		// setup
-		setInternalState(decoraTagVisitor, "decoraTagName", "TEST");
+		decoraTagVisitor.decoraTagName = "TEST";
 
 		// when
-		invokeMethod(decoraTagVisitor, "checkNestedDecoraTags");
 
-		// then
-		fail("A DecoraException must have occured because decoraTagName is not null.");
+		assertThrows(DecoraException.class, () -> {
+			decoraTagVisitor.checkNestedDecoraTags();
+		});
 	}
 
 	@Test
 	public final void testCheckNestedDecoraTagsDecoraTagNameNull() throws Exception {
 		// setup
-		String nullString = null;
-		setInternalState(decoraTagVisitor, "decoraTagName", nullString);
+		decoraTagVisitor.decoraTagName = null;
 
 		// when
-		invokeMethod(decoraTagVisitor, "checkNestedDecoraTags");
+		decoraTagVisitor.checkNestedDecoraTags();
 
 		// then
 		// DecoraException not expected