+++ jodd-http/src/test/java/jodd/http/HttpBrowserOfflineTest.java
@@ -25,9 +25,9 @@
 
 package jodd.http;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class HttpBrowserOfflineTest {
 
@@ -41,7 +41,7 @@ public class HttpBrowserOfflineTest {
 
 		httpBrowser.addDefaultHeaders(request);
 
-		assertEquals(3, request.headers().size());
+		assertEquals(3, request.headerNames().size());
 		assertEquals("123", request.header("aaa"));
 		assertEquals("987", request.header("bbb"));
 	}
@@ -56,7 +56,7 @@ public class HttpBrowserOfflineTest {
 
 		httpBrowser.addDefaultHeaders(request);
 
-		assertEquals(2, request.headers().size());
+		assertEquals(2, request.headerNames().size());
 		assertEquals("987", request.header("aaa"));
 	}
 }