+++ jodd-http/src/test/java/jodd/http/HttpMultiMapTest.java
@@ -25,23 +25,23 @@
 
 package jodd.http;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
 
 public class HttpMultiMapTest {
 
 	@Test
 	public void testAdd() {
-		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitiveMap();
 
 		mm.add("One", "one");
 		mm.add("Two", "two");
@@ -53,7 +53,7 @@ public class HttpMultiMapTest {
 
 	@Test
 	public void testAddSameName() {
-		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitiveMap();
 
 		mm.add("One", "one");
 		mm.add("one", "two");
@@ -76,14 +76,14 @@ public class HttpMultiMapTest {
 
 	@Test
 	public void testMissing() {
-		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitiveMap();
 
 		assertNull(mm.get("xxx"));
 	}
 
 	@Test
 	public void testIterator() {
-		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitiveMap();
 
 		mm.add("One", "one");
 		mm.add("one", "two");
@@ -103,7 +103,7 @@ public class HttpMultiMapTest {
 
 		try {
 			i.next();
-			fail();
+			fail("error");
 		} catch (Exception ignore) {}
 
 		mm.clear();
@@ -114,7 +114,7 @@ public class HttpMultiMapTest {
 
 	@Test
 	public void testNullValues() {
-		HttpMultiMap<String> hmm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> hmm = HttpMultiMap.newCaseInsensitiveMap();
 
 		assertFalse(hmm.contains("one"));
 
@@ -136,7 +136,7 @@ public class HttpMultiMapTest {
 
 	@Test
 	public void testParametersNumber() {
-		HttpMultiMap<String> hmm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> hmm = HttpMultiMap.newCaseInsensitiveMap();
 
 		for (int i = 0; i < 30; i++) {
 			hmm.add(String.valueOf(i), "!" + i);
@@ -147,7 +147,7 @@ public class HttpMultiMapTest {
 
 	@Test
 	public void testLetterCaseInsensitive() {
-		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitveMap();
+		HttpMultiMap<String> mm = HttpMultiMap.newCaseInsensitiveMap();
 
 		mm.add("one", "1.1");
 		mm.add("one", "1.1.1");
@@ -169,7 +169,7 @@ public class HttpMultiMapTest {
 
 	@Test
 	public void testLetterCaseSensitive() {
-		HttpMultiMap<String> mm = HttpMultiMap.newCaseSensitveMap();
+		HttpMultiMap<String> mm = HttpMultiMap.newCaseSensitiveMap();
 
 		mm.add("one", "1.1");
 		mm.add("one", "1.1.1");