+++ jodd-vtor/src/test/java/jodd/vtor/constraint/EqualToFieldConstraintTest.java
@@ -27,17 +27,16 @@ package jodd.vtor.constraint;
 
 import jodd.vtor.ValidationConstraintContext;
 import jodd.vtor.VtorException;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.stub;
+import static org.mockito.Mockito.when;
 
 public class EqualToFieldConstraintTest extends ConstraintTestBase {
     @Test
     public void testValidate_withNullValue() {
-        assertTrue("result must be true when validate null value",
-                EqualToFieldConstraint.validate(new Object(), null, "someField"));
+        assertTrue(EqualToFieldConstraint.validate(new Object(), null, "someField"));
     }
 
     @Test
@@ -50,7 +49,7 @@ public class EqualToFieldConstraintTest extends ConstraintTestBase {
     public void testConstructor2() {
         String fieldName = "testField";
         EqualToFieldConstraint equalToFieldConstraint = new EqualToFieldConstraint(fieldName);
-        assertEquals("field name must be the same as was given to constructor", equalToFieldConstraint.getFieldName(), fieldName);
+        assertEquals(fieldName, equalToFieldConstraint.getFieldName());
     }
 
     @Test
@@ -58,7 +57,7 @@ public class EqualToFieldConstraintTest extends ConstraintTestBase {
         EqualToFieldConstraint equalToFieldConstraint = new EqualToFieldConstraint();
         String fieldName = "someField";
         equalToFieldConstraint.setFieldName(fieldName);
-        assertEquals("value must be the same as was given to set method", equalToFieldConstraint.getFieldName(), fieldName);
+        assertEquals(fieldName, equalToFieldConstraint.getFieldName());
     }
 
     @Test
@@ -67,41 +66,39 @@ public class EqualToFieldConstraintTest extends ConstraintTestBase {
         //set a field name through an annotation
         EqualToField fldAnnotation = mock(EqualToField.class);
         String fieldName = "anotherField";
-        stub(fldAnnotation.value()).toReturn(fieldName);
+        when(fldAnnotation.value()).thenReturn(fieldName);
 
         equalToFieldConstraint.configure(fldAnnotation);
-        assertEquals("field name must be the same as was set to annotation when configure",
-                equalToFieldConstraint.getFieldName(), fieldName);
+        assertEquals(fieldName, equalToFieldConstraint.getFieldName());
     }
 
     @Test
     public void testIsValid_forEqualValues() {
         EqualToFieldConstraint equalToDeclaredFieldConstraint = new EqualToFieldConstraint("testField");
         ValidationConstraintContext cvv = mockContext();
-        stub(cvv.getTarget()).toReturn(new TestBean("someValue"));
+        when(cvv.getTarget()).thenReturn(new TestBean("someValue"));
 
-        assertTrue("result must be true when field and value are equals", equalToDeclaredFieldConstraint.isValid(cvv, "someValue"));
+        assertTrue(equalToDeclaredFieldConstraint.isValid(cvv, "someValue"));
     }
 
     @Test
     public void testIsValid_forDifferentValues() {
         EqualToFieldConstraint equalToDeclaredFieldConstraint = new EqualToFieldConstraint("testField");
         ValidationConstraintContext cvv = mockContext();
-        stub(cvv.getTarget()).toReturn(new TestBean("someValue"));
-        assertFalse("result must be false when validated field and value are different", equalToDeclaredFieldConstraint.isValid(cvv, "wrongValue"));
+        when(cvv.getTarget()).thenReturn(new TestBean("someValue"));
+        assertFalse(equalToDeclaredFieldConstraint.isValid(cvv, "wrongValue"));
     }
 
-    @Test(expected = VtorException.class)
+    @Test
     public void testValidate_FieldNotFound() {
         TestBean testVal = new TestBean("someValue");
-        EqualToFieldConstraint.validate(testVal, "someValue", "wrongField");
-        fail("EqualToFieldConstraint should throw VtorException when receive nonexistent field name");
+        assertThrows(VtorException.class, () -> EqualToFieldConstraint.validate(testVal, "someValue", "wrongField"));
     }
 
     @Test
     public void testValidate_FieldValueIsNull() {
         TestBean testVal = new TestBean(null);
-        assertFalse("result must be false when field value is null", EqualToFieldConstraint.validate(testVal, "someValue", "testField"));
+        assertFalse(EqualToFieldConstraint.validate(testVal, "someValue", "testField"));
     }
 
     public static class TestBean {